refactor: entities/features type 분리
