+++ src/entities/comment/model/types.ts
@@ -22,25 +22,6 @@ export interface CommentUserType {
 // 댓글 조회
 export type GetCommentsResponseType = CommentsType;
 
-// 댓글 추가
-export interface PostCommentRequestType {
-  body: string;
-  postId: number | null;
-  userId: number;
-}
-export interface PostCommentResponseType {
-  id: number;
-  body: string;
-  postId: number;
-  user: CommentUserType;
-}
-
-// 댓글 수정
-export interface PutCommentRequestType {
-  body: string;
-}
-export type PutCommentResponseType = CommentType;
-
 // 댓글 좋아요
 export interface PatchCommentLikeRequestType {
   like: number;
