+++ src/pages/PostsManagerPage.tsx
@@ -25,7 +25,8 @@ import {
   TableRow,
   Textarea,
 } from "../shared/ui";
-import { CommentType, PostCommentRequestType } from "@/entities/comment/model/types";
+import { PostCommentRequestType } from "@/features/comment/model/types";
+import { CommentType } from "@/entities/comment/model/types";
 import { GetPostsResponseType, PostType, TagType } from "@/entities/post/model/types";
 import { GetUsersResponseType, UserDetailType, UserType } from "@/entities/user/model/types";
